pyproject.toml
[tool.poetry]
name = "photo2geo"
version = "0.1.0"
description = ""
authors = ["Hi-king <hikingko1@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
requests = "^2.26.0"
pillow = "^8.4.0"
numpy = "^1.21.4"
pandas = "*"
fire = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
